REF: move coordinate-only helper and functions from maputils to coordinates.utils
